backend/requirements.txt
fastapi
uvicorn
pytesseract
pdfplumber
pymupdf
Pillow
chromadb
openai
langchain
python-dotenv
python-multipart